Hero: pulled Hero code into its own component
